requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
# Deep learning
torch>=2.0.0
# Bioinformatics
biopython>=1.81
# Data handling
pyarrow>=14.0.0
requests>=2.31.0
# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
# Configuration
pyyaml>=6.0
# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
# Code quality
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
# Utilities
tqdm>=4.66.0
joblib>=1.3.0
